Totally revamped draft - NOT proofread
Realised the initial idea to write about the prallel in general terms
required too much thinking. So did a quick and dirty CS-oriented
write-up instead. This contains the core set of ideas that can be
expanded into an more general-purpose explanation accessible, at least
to all engineering design people in the next revision. You are welcome
to fork this one and DIY.