My first attempt at the Miller-Rabin modification
The fermat-test procedure fails because the expmod statements, in the
nested conditional, do not evaluate to a value (nested under the 'even?'
conditional). This is because the procedure attempts to check an
un-computed expression against '1' and '(- m 1)' fail. However, the
sequence of checks appears to be correct. Must find some other way to do
this.