MY LOGIC WORKED THE FIRST TIME!!!!!
SO GLAD I did NOT peek :))) The trick was to figure out the
substitutions. I couldn't see it for the longest time... After many
paper designs, I stumbled upon this approach that just made more sense
with each step. Now I cannot believe that it just worked. Time for a
beer :)