Replace simpsons-rule integration procedure
Use the rule:
Numerical Integration of a function f over limit a-to-b, in increments
of dx =
[f(a + dx/2) + f(a + dx + dx/2) +  f(a + 2dx + dx/2) + ... ]*dx

This enables two things:
1. A far more *clear* *demonstration* of the broad applicability of
'accumulate'. The simpsons-rule procedure has too much cruft to properly
serve this demo purpose.
2. Is expressible using primitives that conforms to the latest latest
specification of the Scheme language, R6RS. The online REPL I have
suggested for testing the script is being built to conform to this
standard. It refuses to evaluate my simpsons-rule code, as I have used
the 'remainder' primitive for one of the internal definitions.
Apparently, this primitive is no longer supported by the latest Scheme
language standard.

Lastly, I tweaked some of the test arguments and also added expected
values to look for, if the suggested arguments are used.