Fixed if-condition in count-frac-iter
Testing for count '<=' 2 was plain wrong... it should be just count '='
1. This is because the loop is invoked with an already-decremented
value. To be specific, if the current value of (count - 1) in the loop
is 2, then it must compute d (2 - 1) +  n(2)/result. i.e. the
if-condition must allow the result to be transformed as long as count is
2 or more, and should terminate only when the countdown value reaches 1.

This bug was revealed when I used the (buggy) cont-frac-iter procedure
to compute Euler's constant e, as asked in ex.1.38.