Defined numbers one and two as pure functions!
It took a while to walk through the evaluation of (add-1 zero) and
(add-1 one by hand). I was confused by how to deal with the nested
lambdas for the longest time. Sorted after a few experiments in REPL and
re-re-re-readings of the introduced to lambda procedures on page 63 of
the book.

Now working our how to define + in pure functional terms without using
scalar numerals.