Use 'racket', not 'sicp' in ex2-44 picture lang file
As of this commit, the code reliably always runs with #lang racket,
and always fails with #lang sicp.

Failure emits this contract violation:

    segments->painter: contract violation
      expected: list?
      given: (#(struct:segment #(struct:vect 0 0) #(struct:vect 0.99 0))
    #(struct:segment #(struct:vect 0.99 0) #(struct:vect 0.99 0.99))
    #(struct:segment #(struct:vect 0.99 0.99) #(struct:vect 0 0.99))
    #(struct:segment #(struct:vect 0 0.99) #(struct:vect 0 0)))
      in: the 1st argument of
          {-> {listof any/c} any/c}
      contract from:
          <pkgs>/sicp/sicp-pict/main.rkt
      blaming:
    /home/adi/src/github/adityaathalye/sicp/ex2-44-picture-language.rkt
       (assuming the contract is correct)
      at: <pkgs>/sicp/sicp-pict/main.rkt:377.24

I currently have:

    DrRacket version : 7.5
    sicp checksum    : dce62d44790e731b00d53c5acb57aa4410a674a3
                       (installed from the DrRacket package manager)

WHY?

It seems from https://github.com/sicp-lang/sicp/issues/4 that

  - '#lang sicp' constructs lists with _mutable_ cons cells,
  - which violate racket's contract for segments->painter,
  - which contract presumes lists are constructed with racket-land's
    _immutable_cons cells.

The error message does not provide any insight into this subtlety. I
don't know if it can or should.

sicp/issue/4 notes that it was fixed as per SHA 0da6a81, sometime in
March 2016. However it is breaking my code as of this commit.

Said code worked fine as of my prior commit on Feb 19, 2018, with
'#lang sicp':
https://github.com/adityaathalye/sicp/commit/5bbc30e695e7ecb00ff870cbd41b0889d92fe8df

I don't remember the DrRacket version or sicp checksum at that time,
unfortunately. However it's safe to presume the version was current as
of Feb 2018, because I installed DrRacket and sicp package specifically
to see the images being rendered.